requirements.txt
#Web Framework & Environment
flask
gunicorn
python-dotenv

#LangChain & OpenAI Core
langchain
langchain-core
langchain-community
langchain-openai
openai
transformers
sentencepiece
tokenizers

#Vector Store
langchain-chroma
chromadb

#Document Processing
PyMuPDF
pdf2image
Pillow
python-pptx
python-docx
openpyxl

#Google Cloud & Utilities
google-cloud-storage
google-auth
tqdm
langchain-experimental
google-cloud-secret-manager
